Commit #2: test for successful login and unsuccessful login (empty username)
