+++ playwright-workshop/src/test/java/com/devexperts/in/qatesting/configuration/LoginTest.java
@@ -1,4 +1,4 @@
-package com.devexperts.in.qatesting;
+package com.devexperts.in.qatesting.configuration;
 
 import com.microsoft.playwright.*;
 import com.microsoft.playwright.options.AriaRole;
@@ -13,7 +13,7 @@ public class LoginTest {
     public void testSuccessfulLogin(){
 
         Playwright playwright = Playwright.create();
-        Browser browser = playwright.chromium().launch(new BrowserType.LaunchOptions().setHeadless(false));
+        Browser browser = playwright.chromium().launch();
         Page page = browser.newPage();
         page.navigate("https://qa-testing.in.devexperts.com/internship/");
         Locator inputUsername = page.getByPlaceholder("Username");
@@ -27,6 +27,31 @@ public class LoginTest {
                 () ->assertThat(homeHeaderPage).hasText("Home Test Task"),
                 () ->assertThat(homeHeaderPage).isVisible());
 
+        page.close();
+        browser.close();
+        playwright.close();
+    }
+
+    @Test
+    public void testUnsuccessfulLoginWithBlankUsername(){
+        //setup: Create playwrightinstance, browser, page
+        Playwright playwright = Playwright.create();
+        Browser browser = playwright.chromium().launch();
+        Page page = browser.newPage();
+        //open login page
+        page.navigate("https://qa-testing.in.devexperts.com/internship/");
+        //inform th username, blank
+        // pass
+        Locator inputPassword = page.locator("#password");
+        inputPassword.fill("Test1234!");
+        //click login button
+        Locator loginButton = page.getByRole(AriaRole.BUTTON, new Page.GetByRoleOptions().setName("Login"));
+        loginButton.click();
+        //  check if we were not redirected to the homepage
+        //login-status
+        Locator errorText = page.locator("#login-status");
+        assertThat(errorText).hasText("Please enter valid credentials:");
+        // close everything
         page.close();
         browser.close();
         playwright.close();
